refactor: split transformer support helpers
