+++ crates/bitnet-transformer/src/lib.rs
@@ -1,7 +1,6 @@
-use bitnet_common::{
-    BitNetConfig, BitNetError, Result,
-    config::{ActivationType, NormType},
-};
+#[cfg(test)]
+use bitnet_common::config::NormType;
+use bitnet_common::{BitNetConfig, BitNetError, Result, config::ActivationType};
 use bitnet_qk256_dispatch::{
     forward_qk256_with_scale, record_bitnet_linear_cpu_fallback, record_bitnet_linear_unsupported,
     strict_cuda_bitnet_backend_requested,
@@ -9,380 +8,23 @@ use bitnet_qk256_dispatch::{
 use bitnet_rope::{build_tables as build_rope_tables, resolve_base as resolve_rope_base};
 use candle_core::{DType, Device, Module, Tensor};
 use candle_nn::{LayerNorm, Linear, VarBuilder};
-use std::sync::OnceLock;
-
 mod attention_forward;
 
-#[derive(Clone, Debug)]
-struct QwenTraceConfig {
-    path: Option<std::path::PathBuf>,
-    stderr_enabled: bool,
-    layer: usize,
-}
+mod diagnostics;
+mod layer_builders;
+mod qk256;
 
-fn env_flag(name: &str) -> bool {
-    std::env::var(name).is_ok()
-}
-
-fn env_flag_eq_1(name: &str) -> bool {
-    std::env::var(name).as_deref() == Ok("1")
-}
-
-fn qwen_trace_config() -> &'static QwenTraceConfig {
-    static CONFIG: OnceLock<QwenTraceConfig> = OnceLock::new();
-    CONFIG.get_or_init(|| QwenTraceConfig {
-        path: std::env::var("BITNET_QWEN_TRACE_JSONL")
-            .ok()
-            .filter(|value| !value.trim().is_empty())
-            .map(std::path::PathBuf::from),
-        stderr_enabled: env_flag_eq_1("BITNET_QWEN_TRACE"),
-        layer: std::env::var("BITNET_QWEN_TRACE_LAYER")
-            .ok()
-            .and_then(|value| value.parse::<usize>().ok())
-            .unwrap_or(0),
-    })
-}
-
-fn debug_attn_enabled() -> bool {
-    static ENABLED: OnceLock<bool> = OnceLock::new();
-    *ENABLED.get_or_init(|| env_flag("DEBUG_ATTN"))
-}
-
-fn debug_attn_scale_enabled() -> bool {
-    static ENABLED: OnceLock<bool> = OnceLock::new();
-    *ENABLED.get_or_init(|| env_flag("BITNET_DEBUG_ATTN_SCALE"))
-}
-
-fn debug_gqa_enabled() -> bool {
-    static ENABLED: OnceLock<bool> = OnceLock::new();
-    *ENABLED.get_or_init(|| env_flag("BITNET_DEBUG_GQA"))
-}
-
-fn debug_rope_enabled() -> bool {
-    static ENABLED: OnceLock<bool> = OnceLock::new();
-    *ENABLED.get_or_init(|| env_flag("BITNET_DEBUG_ROPE"))
-}
-
-fn debug_rmsnorm_enabled() -> bool {
-    static ENABLED: OnceLock<bool> = OnceLock::new();
-    *ENABLED.get_or_init(|| env_flag("BITNET_DEBUG_RMSNORM"))
-}
-
-fn debug_mlp_enabled() -> bool {
-    static ENABLED: OnceLock<bool> = OnceLock::new();
-    *ENABLED.get_or_init(|| env_flag("BITNET_DEBUG_MLP"))
-}
-
-fn trace_rms_enabled() -> bool {
-    static ENABLED: OnceLock<bool> = OnceLock::new();
-    *ENABLED.get_or_init(|| env_flag_eq_1("BITNET_TRACE_RMS"))
-}
-
-fn qwen_trace_path() -> Option<std::path::PathBuf> {
-    qwen_trace_config().path.clone()
-}
-
-fn qwen_trace_enabled() -> bool {
-    let config = qwen_trace_config();
-    config.path.is_some() || config.stderr_enabled
-}
-
-fn qwen_trace_active() -> bool {
-    qwen_trace_enabled() && std::env::var("BITNET_QWEN_TRACE_ACTIVE").as_deref() == Ok("1")
-}
-
-fn qwen_trace_layer_enabled(layer_idx: usize) -> bool {
-    if !qwen_trace_active() {
-        return false;
-    }
-    qwen_trace_config().layer == layer_idx
-}
-
-fn qwen_trace_escape(value: &str) -> String {
-    value.replace('\\', "\\\\").replace('"', "\\\"").replace('\n', "\\n").replace('\r', "\\r")
-}
-
-fn qwen_trace_number(value: f64) -> String {
-    if value.is_finite() { format!("{value:.9}") } else { "null".to_string() }
-}
-
-fn qwen_trace_write_line(line: &str) {
-    if let Some(path) = qwen_trace_path() {
-        if let Some(parent) = path.parent()
-            && let Err(err) = std::fs::create_dir_all(parent)
-        {
-            eprintln!("qwen_trace_write_failed: create_dir_all {}: {err}", parent.display());
-            return;
-        }
-        match std::fs::OpenOptions::new().create(true).append(true).open(&path) {
-            Ok(mut file) => {
-                if let Err(err) = std::io::Write::write_all(&mut file, line.as_bytes())
-                    .and_then(|_| std::io::Write::write_all(&mut file, b"\n"))
-                {
-                    eprintln!("qwen_trace_write_failed: {}: {err}", path.display());
-                }
-            }
-            Err(err) => eprintln!("qwen_trace_write_failed: {}: {err}", path.display()),
-        }
-    } else if std::env::var("BITNET_QWEN_TRACE").as_deref() == Ok("1") {
-        eprintln!("{line}");
-    }
-}
-
-fn qwen_trace_event(stage: &str, fields_json: &str) {
-    if !qwen_trace_enabled() {
-        return;
-    }
-    let step = std::env::var("BITNET_QWEN_TRACE_STEP").unwrap_or_else(|_| "null".to_string());
-    qwen_trace_write_line(&format!(
-        "{{\"kind\":\"qwen_trace_event\",\"stage\":\"{}\",\"step\":{},{} }}",
-        qwen_trace_escape(stage),
-        step,
-        fields_json
-    ));
-}
-
-fn qwen_trace_tensor(
-    stage: &str,
-    layer_idx: Option<usize>,
-    tensor: &Tensor,
-) -> candle_core::Result<()> {
-    if !qwen_trace_active() {
-        return Ok(());
-    }
-    if let Some(layer_idx) = layer_idx
-        && !qwen_trace_layer_enabled(layer_idx)
-    {
-        return Ok(());
-    }
-
-    let tensor_f32 =
-        if tensor.dtype() == DType::F32 { tensor.clone() } else { tensor.to_dtype(DType::F32)? };
-    let values = tensor_f32.flatten_all()?.to_vec1::<f32>()?;
-    let mut finite_count = 0usize;
-    let mut nonfinite_count = 0usize;
-    let mut sum = 0.0f64;
-    let mut sum_sq = 0.0f64;
-    let mut min = f64::INFINITY;
-    let mut max = f64::NEG_INFINITY;
-    let mut checksum = 0.0f64;
-    for (idx, value) in values.iter().enumerate() {
-        let value = *value as f64;
-        if value.is_finite() {
-            finite_count += 1;
-            sum += value;
-            sum_sq += value * value;
-            min = min.min(value);
-            max = max.max(value);
-            if idx < 4096 {
-                checksum += value * ((idx % 257) + 1) as f64;
-            }
-        } else {
-            nonfinite_count += 1;
-        }
-    }
-
-    let denom = finite_count.max(1) as f64;
-    let mean = sum / denom;
-    let rms = (sum_sq / denom).sqrt();
-    let sample = values
-        .iter()
-        .take(8)
-        .map(|value| qwen_trace_number(*value as f64))
-        .collect::<Vec<_>>()
-        .join(",");
-    let dims = tensor.dims().iter().map(|dim| dim.to_string()).collect::<Vec<_>>().join(",");
-    let layer_json = layer_idx.map(|idx| idx.to_string()).unwrap_or_else(|| "null".to_string());
-    let step = std::env::var("BITNET_QWEN_TRACE_STEP").unwrap_or_else(|_| "null".to_string());
-
-    qwen_trace_write_line(&format!(
-        "{{\"kind\":\"qwen_trace_tensor\",\"stage\":\"{}\",\"step\":{},\"layer\":{},\"dtype\":\"{:?}\",\"dims\":[{}],\"len\":{},\"finite\":{},\"nonfinite\":{},\"mean\":{},\"rms\":{},\"min\":{},\"max\":{},\"checksum\":{},\"sample\":[{}]}}",
-        qwen_trace_escape(stage),
-        step,
-        layer_json,
-        tensor.dtype(),
-        dims,
-        values.len(),
-        finite_count,
-        nonfinite_count,
-        qwen_trace_number(mean),
-        qwen_trace_number(rms),
-        qwen_trace_number(min),
-        qwen_trace_number(max),
-        qwen_trace_number(checksum),
-        sample
-    ));
-    Ok(())
-}
-
-/// Debug helper for tensor statistics (only runs if DEBUG_ATTN env var is set)
-fn dbg_stats(tag: &str, t: &Tensor) -> candle_core::Result<()> {
-    if debug_attn_enabled() {
-        let mean = t.mean_all()?.to_scalar::<f32>()?;
-        // Compute std manually: sqrt(E[(x - mean)^2])
-        let diff = t.broadcast_sub(&t.mean_all()?)?;
-        let variance = diff.sqr()?.mean_all()?;
-        let std = variance.sqrt()?.to_scalar::<f32>()?;
-        eprintln!("[dbg] {tag}: mean={mean:.6} std={std:.6}");
-    }
-    Ok(())
-}
-
-/// Debug helper for checking finite values
-fn dbg_finite(tag: &str, t: &Tensor) -> candle_core::Result<()> {
-    if debug_attn_enabled() {
-        let v: Vec<f32> = t.flatten_all()?.to_vec1()?;
-        let n = v.len().min(4096);
-        let mut n_nan = 0;
-        let mut n_inf = 0;
-        for &x in &v[..n] {
-            if !x.is_finite() {
-                if x.is_nan() {
-                    n_nan += 1;
-                } else {
-                    n_inf += 1;
-                }
-            }
-        }
-        if n_nan + n_inf > 0 {
-            eprintln!(
-                "⚠️  [dbg] {tag}: non-finite values: NaN={n_nan} Inf={n_inf} (in first {n} elems)"
-            );
-        }
-    }
-    Ok(())
-}
-
-/// Helper to create linear layers with optional bias tensors.
-fn linear_with_optional_bias(
-    in_dim: usize,
-    out_dim: usize,
-    vb: VarBuilder,
-) -> candle_core::Result<Linear> {
-    let weight = vb.get((out_dim, in_dim), "weight")?;
-
-    // Missing bias is semantically a no-bias linear layer. Avoid materializing
-    // zero tensors and a runtime add for dense GGUF weights that omit bias.
-    let bias = match vb.get(out_dim, "bias") {
-        Ok(b) => Some(b),
-        Err(_) => {
-            tracing::debug!("Bias tensor missing for linear layer; using no-bias path [{out_dim}]");
-            None
-        }
-    };
-
-    Ok(Linear::new(weight, bias))
-}
-
-/// Helper to create layer norm with optional bias.
-/// If `bias` is missing we use no-bias LayerNorm by default, or error when
-/// `BITNET_REQUIRE_LAYER_NORM_BIAS=1`.
+use diagnostics::{
+    dbg_finite, dbg_stats, debug_attn_enabled, debug_attn_scale_enabled, debug_gqa_enabled,
+    debug_mlp_enabled, debug_rmsnorm_enabled, debug_rope_enabled, qwen_trace_event,
+    qwen_trace_layer_enabled, qwen_trace_number, qwen_trace_tensor, trace_rms_enabled,
+};
 #[cfg(test)]
-fn layer_norm_with_optional_bias(
-    normalized_shape: usize,
-    eps: f64,
-    vb: VarBuilder,
-) -> candle_core::Result<LayerNorm> {
-    norm_with_optional_bias(NormType::LayerNorm, normalized_shape, eps, vb)
-}
-
-fn norm_with_optional_bias(
-    norm_type: NormType,
-    normalized_shape: usize,
-    eps: f64,
-    vb: VarBuilder,
-) -> candle_core::Result<LayerNorm> {
-    let weight = vb.get((normalized_shape,), "weight")?;
-    if matches!(norm_type, NormType::RmsNorm) {
-        if vb.get((normalized_shape,), "bias").is_ok() {
-            tracing::debug!(
-                "Bias tensor present for RMSNorm layer; ignoring bias [{}]",
-                normalized_shape
-            );
-        }
-        tracing::debug!("Using RMSNorm without mean subtraction [{}]", normalized_shape);
-        return Ok(LayerNorm::rms_norm(weight, eps));
-    }
-
-    match vb.get((normalized_shape,), "bias") {
-        Ok(bias) => {
-            // Bias exists → standard LayerNorm (with mean subtraction and bias)
-            tracing::debug!("Using LayerNorm with bias [{}]", normalized_shape);
-            Ok(LayerNorm::new(weight, bias, eps))
-        }
-        Err(err) => {
-            if std::env::var("BITNET_REQUIRE_LAYER_NORM_BIAS")
-                .ok()
-                .is_some_and(|value| value == "1")
-            {
-                return Err(candle_core::Error::Msg(format!(
-                    "LayerNorm bias tensor is required but missing (set BITNET_REQUIRE_LAYER_NORM_BIAS=0 or unset to allow no-bias LayerNorm): {err}"
-                )));
-            }
-
-            // No bias → LayerNorm without bias (but WITH mean subtraction)
-            // IMPORTANT: Use LayerNorm::new_no_bias (remove_mean=true) NOT rms_norm (remove_mean=false)
-            // because these gamma weights are calibrated for LayerNorm semantics
-            // (mean subtraction). RMSNorm callers return earlier in this helper.
-            tracing::debug!(
-                "Bias tensor missing for norm layer; using LayerNorm without bias (mean subtraction enabled) [{}]",
-                normalized_shape
-            );
-            Ok(LayerNorm::new_no_bias(weight, eps))
-        }
-    }
-}
-
-fn optional_layer_norm_with_optional_bias(
-    norm_type: NormType,
-    normalized_shape: usize,
-    eps: f64,
-    vb: VarBuilder,
-) -> candle_core::Result<Option<LayerNorm>> {
-    if !vb.contains_tensor("weight") {
-        return Ok(None);
-    }
-
-    Ok(Some(norm_with_optional_bias(norm_type, normalized_shape, eps, vb)?))
-}
-
-fn qk256_scale_key(qk256_key: &str) -> String {
-    if let Some(base) = qk256_key.strip_suffix(".qk256_qs") {
-        format!("{base}.qk256_scale")
-    } else {
-        format!("{qk256_key}.qk256_scale")
-    }
-}
-
-fn qk256_inline_scale(
-    raw_tensors: &std::collections::HashMap<String, Tensor>,
-    qk256_key: &str,
-) -> Result<Option<f32>> {
-    let scale_key = qk256_scale_key(qk256_key);
-    let Some(scale_tensor) = raw_tensors.get(&scale_key) else {
-        return Ok(None);
-    };
-
-    let scale_values = scale_tensor.flatten_all()?.to_vec1::<f32>().map_err(|e| {
-        BitNetError::Validation(format!("failed to read QK256 inline scale {scale_key}: {e}"))
-    })?;
-    let [scale] = scale_values.as_slice() else {
-        return Err(BitNetError::Validation(format!(
-            "QK256 inline scale {scale_key} must contain exactly one value, got {}",
-            scale_values.len()
-        )));
-    };
-    let scale = *scale;
-    if !scale.is_finite() {
-        return Err(BitNetError::Validation(format!(
-            "QK256 inline scale {scale_key} is not finite: {scale}"
-        )));
-    }
-
-    Ok(Some(scale))
-}
-
-const TIED_EMBED_QK256_KEY: &str = "embed_tokens.weight.qk256_qs";
+use layer_builders::layer_norm_with_optional_bias;
+use layer_builders::{
+    linear_with_optional_bias, norm_with_optional_bias, optional_layer_norm_with_optional_bias,
+};
+use qk256::{TIED_EMBED_QK256_KEY, qk256_inline_scale};
 
 /// Rotary Position Embedding
 pub struct RotaryEmbedding {
